+++ src/Indice.Features.Cases.App/src/app/features/case-types/case-types.component.ts
@@ -1,9 +1,10 @@
 import { map, take } from 'rxjs/operators';
-import { CasesApiService, CaseTypePartial, CaseTypePartialResultSet } from './../../core/services/cases-api.service';
 import { Component, OnInit } from '@angular/core';
 import { BaseListComponent, Icons, IResultSet, RouterViewAction, ViewAction, ListViewType, ModalService } from '@indice/ng-components';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Observable } from 'rxjs';
+import { CaseTypeService } from 'src/app/core/services/case-type.service';
+import { CaseTypePartial, CasesApiService, CaseTypePartialResultSet } from 'src/app/core/services/cases-api.service';
 import { CaseTypeDeleteModalComponent } from './case-type-delete-modal/case-type-delete-modal.component';
 
 @Component({
@@ -21,7 +22,8 @@ export class CaseTypesComponent extends BaseListComponent<CaseTypePartial> imple
     private _api: CasesApiService,
     private _route: ActivatedRoute,
     private _router: Router,
-    private ModalService: ModalService
+    private ModalService: ModalService,
+    private caseTypeService: CaseTypeService
   ) {
     super(_route, _router)
     this.view = ListViewType.Table
@@ -32,7 +34,7 @@ export class CaseTypesComponent extends BaseListComponent<CaseTypePartial> imple
   }
 
   loadItems(): Observable<IResultSet<CaseTypePartial> | null | undefined> {
-    return this._api.getCaseTypes()
+    return this.caseTypeService.getCaseTypes()
       .pipe(
         take(1),
         map((result: CaseTypePartialResultSet) => (result as IResultSet<CaseTypePartial>))
@@ -45,7 +47,7 @@ export class CaseTypesComponent extends BaseListComponent<CaseTypePartial> imple
       keyboard: false,
       initialState: {id: caseTypeId}
     });
-    
+
     modal.onHidden?.subscribe(_ => {
       if ((_ as any).result) {
         this.refresh();