+++ src/Indice.Features.Media.AspNetCore/MediaLibraryFeatureExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using Indice.Features.Media.AspNetCore;
 using Indice.Features.Media.AspNetCore.Authorization;
+using Indice.Features.Media.AspNetCore.Events;
 using Indice.Features.Media.AspNetCore.Services;
 using Indice.Features.Media.AspNetCore.Services.Hosting;
 using Indice.Features.Media.AspNetCore.Stores;
@@ -48,10 +49,13 @@ public static class MediaLibraryFeatureExtensions
             services.AddHostedService<FoldersCleanUpHostedService>();
             services.AddHostedService<FilesCleanUpHostedService>();
         }
+        services.AddBackgroundPlatformEventService();
+        services.AddPlatformEventHandler<FolderRenameCommand, FolderRenameCommandHandler>();
         services.AddSingleton(new DatabaseSchemaNameResolver(apiOptions.DatabaseSchema));
         // Register application DbContext.
         services.AddDbContext<MediaDbContext>(apiOptions.ConfigureDbContext ?? ((serviceProvider, builder) => builder.UseSqlServer(serviceProvider.GetRequiredService<IConfiguration>().GetConnectionString("MediaLibraryDbConnection"))));
 
+        services.AddDistributedMemoryCache();
         // Register Default Policy Provider.
         // Add authorization policies that are used by the IdentityServer API.